Updated generatePermutations
It Generates the permutations slightly more now. Now it finds the max elements first and then it goes from right to left and left to right until all permutations are done. I probably need to tweak it a bit though since it's doing more than necessary and I think there's still bugs that need to be ironed out